exosphere-runtimes/parse-list-of-docs/pyproject.toml
[project]
name = "parse-list-of-docs-runtime"
version = "0.1.0"
description = "Simplified document processing workflow for parsing lists of documents from S3"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exospherehost==0.0.7b3",
    "python-dotenv>=1.1.1",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "pydantic>=1.10.0",
    "aiohttp>=3.8.0",
    "boto3>=1.26.0",
]
